refactor(clip): 폴더 저장 유스케이스 통합
