--- src/folders/application/usecases/get-folders.usecase.ts
@@ -1,17 +1,0 @@
-import { FoldersRepository } from '../../domain/folders.repository';
-
-export class GetFoldersUseCase {
-  constructor(private readonly foldersRepository: FoldersRepository) {}
-
-  async execute(userId: string) {
-    const workspaceId = await this.foldersRepository.findPersonalWorkspaceId(
-      userId,
-    );
-
-    if (!workspaceId) {
-      return [];
-    }
-
-    return this.foldersRepository.findFoldersByWorkspaceId(workspaceId);
-  }
-}
