+++ src/folders/application/usecases/save-folder.usecase.ts
@@ -1,0 +1,50 @@
+import { FoldersRepository } from '../../domain/folders.repository';
+import { FoldersError } from '../folders.error';
+
+export type SaveFolderInput =
+  | {
+      mode: 'create';
+      name: string;
+    }
+  | {
+      mode: 'update';
+      folderId: string;
+      name?: string;
+    };
+
+export class SaveFolderUseCase {
+  constructor(private readonly foldersRepository: FoldersRepository) {}
+
+  async execute(userId: string, input: SaveFolderInput) {
+    if (input.mode === 'create') {
+      const workspaceId =
+        await this.foldersRepository.getOrCreatePersonalWorkspaceId(userId);
+
+      const lastOrder = await this.foldersRepository.findLastFolderOrder(
+        workspaceId,
+      );
+      const nextOrder = lastOrder ? lastOrder + 1 : 1;
+
+      return this.foldersRepository.createFolder({
+        name: input.name,
+        order: nextOrder,
+        workspaceId,
+      });
+    }
+
+    const folder = await this.foldersRepository.findPersonalFolderById(
+      userId,
+      input.folderId,
+    );
+
+    if (!folder) {
+      throw new FoldersError('NOT_FOUND', '폴더를 찾을 수 없습니다.');
+    }
+
+    if (!input.name) {
+      return folder;
+    }
+
+    return this.foldersRepository.updateFolderName(folder.id, input.name);
+  }
+}
